Move Preact10ShallowDiff into compat-shallow-renderer sub directory
