+++ test/Preact10ShallowDiff/Preact10ShallowDiffHooks_test.tsx
@@ -9,7 +9,7 @@
 import type { VNode } from 'preact';
 import * as preact from 'preact/compat';
 import type { StateUpdater } from 'preact/compat';
-import Preact10ShallowDiff from '../../src/Preact10ShallowDiff.js';
+import Preact10ShallowDiff from '../../src/compat-shallow-renderer/Preact10ShallowDiff.js';
 import { expect, installVNodeTestHook } from './utils.js';
 
 const {
