+++ test/Preact10ShallowDiff/Preact10ShallowDiffMemo_test.tsx
@@ -13,7 +13,7 @@ import type { ComponentType, VNode } from 'preact';
 import PropTypes from 'prop-types';
 import sinon from 'sinon';
 
-import Preact10ShallowDiff from '../../src/Preact10ShallowDiff.js';
+import Preact10ShallowDiff from '../../src/compat-shallow-renderer/Preact10ShallowDiff.js';
 import { expect, installVNodeTestHook } from './utils.js';
 
 const {
